% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_constant}
\alias{async_constant}
\title{Make a minimal deferred that resolves to the specified value}
\usage{
async_constant(value = NULL)
}
\arguments{
\item{value}{The value to resolve to.}
}
\value{
A deferred value.
}
\description{
This is sometimes useful to start a deferred chain.
}
\details{
Note that the evaluation of \code{value} is forced when the deferred value
is created.
}
\section{Examples}{
\preformatted{afun <- async(function() {
  async_constant(1/100)$
    then(function(x) delay(x))$
    then(function(x) print(x))
})
synchronise(afun())
}
}

\keyword{internal}
