% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-cache.R
\name{cmc__update_primary}
\alias{cmc__update_primary}
\title{Update the primary cache from the replica cache}
\usage{
cmc__update_primary(self, private, rds, packages, lock)
}
\arguments{
\item{self}{self}

\item{private}{private self}

\item{rds}{Whether to update the RDS file.}

\item{packages}{Whether to update the PACKAGES files (+ ETag files).}
}
\value{
Nothing.
}
\description{
Update the primary cache from the replica cache
}
\keyword{internal}
