% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{default_cran_mirror}
\alias{default_cran_mirror}
\title{Query the default CRAN repository for this session}
\usage{
default_cran_mirror()
}
\value{
A named character vector of length one, where the
name is \code{"CRAN"}.
}
\description{
If \code{options("repos")} (see \code{\link[=options]{options()}}) contains an entry called
\code{"CRAN"}, then that is returned. If it is a list, it is converted
to a character vector.
}
\details{
Otherwise the RStudio CRAN mirror is used.
}
\examples{
default_cran_mirror()
}
