% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_graphics_api_version}
\alias{get_graphics_api_version}
\title{Query the version of the graphics API}
\usage{
get_graphics_api_version()
}
\value{
An integer scalar, the version of the graphics API of this
R version.
}
\description{
A package compiled with a certain version of the graphics API
will not work with R installations that use a different version.
}
\examples{
get_graphics_api_version()
}
