% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_internals_id}
\alias{get_internals_id}
\title{Query UUID identifying the version of the R API}
\usage{
get_internals_id()
}
\value{
String, a UUID.
}
\description{
Packages need to be recompiled if this id changes.
}
\examples{
get_internals_id()
}
