% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-sso.R
\name{ppm_sso_login}
\alias{ppm_sso_login}
\alias{ppm_sso_logout}
\alias{ppm_sso_status}
\title{Posit Package Manager single sign-on (SSO) authentication}
\usage{
ppm_sso_login()

ppm_sso_logout()

ppm_sso_status(connect = FALSE)
}
\arguments{
\item{connect}{If \code{TRUE}, also checks if the token is valid by making a test
request to the Package Manager instance. This requires an active internet
connection and may take a few seconds. If \code{FALSE}, only checks if a
token is cached and not expired.}
}
\description{
Posit Package Manager single sign-on (SSO) authentication
}
\details{
\code{ppm_sso_login()} initiates the SSO login process. You should be
prompted to log in via your browser, and the obtained token will be
cached for future use.
\subsection{Set up SSO authentication:}{
\itemize{
\item Set the \code{PACKAGEMANAGER_ADDRESS} environment variable to the URL of
your RStudio Package Manager instance. For example, add this line to
your \code{.Renviron} file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{PACKAGEMANAGER_ADDRESS=https://<ppm-url>
}\if{html}{\out{</div>}}

Alternatively, you can also set it in your shell profile on Unix,
or in the System or User environment variables on Windows.
\item Set \code{options(repos)} to include a repository from your Package Manager
instance. Include \verb{__token__} as the username in the URL. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(repos = c(
  PPM = "https://__token__@<ppm-url>/<repo-path>",
  getOption("repos")
))
}\if{html}{\out{</div>}}
\item Call \code{\link[=repo_get]{repo_get()}} to trigger authentication and caching of the token.
You should be prompted to log in via your browser, and the obtained
token will be cached for future use. Call \code{\link[=ppm_sso_status]{ppm_sso_status()}} to check
the status of your authentication, including the path of the cached
token and its expiration time.
\item Alternatively, you can call \code{ppm_sso_login()} directly to trigger
the login process directly.
}
}

\code{ppm_sso_logout()} removes the cached token, effectively logging you
out.

\code{ppm_sso_status()} checks the status of your authentication, including
the path of the cached token and its expiration time.
}
\seealso{
\url{https://docs.posit.co/rspm/admin/authentication/}
}
