% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{current_r_platform}
\alias{current_r_platform}
\alias{default_platforms}
\title{Platform and R version information of the current session}
\usage{
current_r_platform()

default_platforms()
}
\value{
\code{current_r_platform()} returns a character scalar.

\code{default_platform()} returns a character vector of the
default platforms.
}
\description{
Currently used platforms:
\itemize{
\item \code{"source"}
\item \code{"macos"}
\item \code{"windows"}
}
}
\examples{
current_r_platform()
default_platforms()
}
