% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{current_r_platform}
\alias{current_r_platform}
\alias{current_r_platform_data}
\alias{default_platforms}
\title{R platforms}
\usage{
current_r_platform()

current_r_platform_data()

default_platforms()
}
\value{
\code{current_r_platform()} returns a character scalar.

\code{current_r_platform_data()} returns a data frame with character
scalar columns:
\itemize{
\item \code{cpu},
\item \code{vendor},
\item \code{os},
\item \code{distribution} (only on Linux),
\item \code{release} (only on Linux),
\item \code{platform}: the concatenation of the other columns, separated by
a dash.
}

\code{default_platforms()} returns a character vector of the
default platforms.
}
\description{
R platforms
}
\details{
\code{current_r_platform()} detects the platform of the current R version.
\code{current_r_platform_data()} is similar, but returns the raw data instead
of a character scalar.

By default pkgcache works with source packages and binary packages for
the current platform. You can change this, by providing different
platform names as arguments to
\code{\link[=cranlike_metadata_cache]{cranlike_metadata_cache$new()}},
\code{\link[=repo_status]{repo_status()}}, etc.

These functions accept the following platform names:
\itemize{
\item \code{"source"} for source packages,
\item \code{"macos"} for macOS binaries that are appropriate for the R versions
pkgcache is working with by dropping packages for incompatible
CPU architectures (defaulting to the CPU of the current machine),
as defined by CRAN binaries. E.g. on R 3.5.0 macOS binaries
are built for macOS El Capitan.
\item \code{"windows"} for Windows binaries for the default CRAN architecture.
This is currently Windows Vista for all supported R versions, but it
might change in the future. The actual binary packages in the
repository might support both 32 bit and 64 builds, or only one of
them. In practice 32-bit only packages are very rare. CRAN builds
before and including R 4.1 have both architectures, from R 4.2 they
are 64 bit only. \code{"windows"} is an alias to \code{i386+x86_64-w64-mingw32}
currently.
\item A platform string like \code{R.version$platform}, but on Linux the name
and version of the distribution are also included. Examples:
\itemize{
\item \code{x86_64-apple-darwin17.0}: macOS High Sierra.
\item \code{aarch64-apple-darwin20}: macOS Big Sur on arm64.
\item \code{x86_64-w64-mingw32}: 64 bit Windows.
\item \code{i386-w64-mingw32}: 32 bit Windows.
\item \code{i386+x86_64-w64-mingw32}: 64 bit + 32 bit Windows.
\item \code{i386-pc-solaris2.10}: 32 bit Solaris. (Some broken 64 Solaris
builds might have the same platform string, unfortunately.)
\item \code{x86_64-pc-linux-gnu-debian-10}: Debian Linux 10 on x86_64.
\item \verb{x86_64-pc-linux-musl-alpine-3.14.1}: Alpine Linux.
\item \code{x86_64-pc-linux-gnu-unknown}: Unknown Linux Distribution on x86_64.
\item \code{s390x-ibm-linux-gnu-ubuntu-20.04}: Ubuntu Linux 20.04 on S390x.
\item \code{amd64-portbld-freebsd12.1}: FreeBSD 12.1 on x86_64.
}
}

\code{default_platfoms()} returns the default platforms for the current R
session. These typically consist of the detected platform of the current
R session, and \code{"source"}, for source packages.
}
\examples{
current_r_platform()
default_platforms()
}
