% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{get_cranlike_metadata_cache}
\alias{get_cranlike_metadata_cache}
\title{The R6 object that implements the global metadata cache}
\usage{
get_cranlike_metadata_cache()
}
\description{
This is used by the \code{\link[=meta_cache_deps]{meta_cache_deps()}}, \code{\link[=meta_cache_list]{meta_cache_list()}}, etc.
functions.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
get_cranlike_metadata_cache()
get_cranlike_metadata_cache()$list("cli")
\dontshow{\}) # examplesIf}
}
