% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pkgcache-package}
\alias{pkgcache}
\alias{pkgcache-package}
\title{Cache for package data and metadata}
\description{
Metadata and package cache for CRAN-like repositories.
     This is a utility package to be used by package management tools
     that want to take advantage of caching.
}
\details{
Metadata and package cache for CRAN-like repositories. This is a utility
package to be used by package management tools that want to take
advantage of caching.
\subsection{Installation}{

You can install the released version of pkgcache from
\href{https://CRAN.R-project.org}{CRAN} with:\if{html}{\out{<div class="sourceCode r">}}\preformatted{install.packages("pkgcache")
}\if{html}{\out{</div>}}
}

\subsection{Metadata cache}{

\code{meta_cache_list()} lists all packages in the metadata cache. It
includes Bioconductor package, and all versions (i.e. both binary and
source) of the packages for the current platform and R version.\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(pkgcache)
meta_cache_list()
#> # A tibble: 42,018 x 32
#>    package      version depends suggests license imports linkingto archs enhances
#>    <chr>        <chr>   <chr>   <chr>    <chr>   <chr>   <chr>     <chr> <chr>   
#>  1 A3           1.0.0   R (>= ~ randomF~ GPL (>~ <NA>    <NA>      <NA>  <NA>    
#>  2 AATtools     0.0.1   R (>= ~ <NA>     GPL-3   magrit~ <NA>      <NA>  <NA>    
#>  3 ABACUS       1.0.0   R (>= ~ rmarkdo~ GPL-3   ggplot~ <NA>      <NA>  <NA>    
#>  4 ABC.RAP      0.9.0   R (>= ~ knitr, ~ GPL-3   graphi~ <NA>      <NA>  <NA>    
#>  5 ABCanalysis  1.2.1   R (>= ~ <NA>     GPL-3   plotrix <NA>      <NA>  <NA>    
#>  6 ABCoptim     0.15.0  <NA>    testtha~ MIT + ~ Rcpp, ~ Rcpp      ABCo~ <NA>    
#>  7 ABCp2        1.2     MASS    <NA>     GPL-2   <NA>    <NA>      <NA>  <NA>    
#>  8 ABHgenotypeR 1.0.1   <NA>    knitr, ~ GPL-3   ggplot~ <NA>      <NA>  <NA>    
#>  9 ABPS         0.3     <NA>    testthat GPL (>~ kernlab <NA>      <NA>  <NA>    
#> 10 ACA          1.1     R (>= ~ <NA>     GPL     graphi~ <NA>      <NA>  <NA>    
#> # ... with 42,008 more rows, and 23 more variables: os_type <chr>,
#> #   priority <chr>, license_is_foss <chr>, license_restricts_use <chr>,
#> #   repodir <chr>, rversion <chr>, platform <list>, needscompilation <chr>,
#> #   ref <chr>, type <chr>, direct <lgl>, status <chr>, target <chr>,
#> #   mirror <chr>, sources <list>, filesize <int>, sha256 <chr>, sysreqs <chr>,
#> #   built <chr>, published <dttm>, deps <list>, md5sum <chr>, path <chr>
}\if{html}{\out{</div>}}

\code{meta_cache_deps()} and \code{meta_cache_revdeps()} can be used to look up
dependencies and reverse dependencies.

The metadata is updated automatically if it is older than seven days,
and it can also be updated manually with \code{meta_cache_update()}.

See the \code{cranlike_metadata_cache} R6 class for a lower level API, and
more control.
}

\subsection{Package cache}{

Package management tools may use the \verb{pkg_cache_*} functions and in
particular the \code{package_cache} class, to make use of local caching of
package files.

The \verb{pkg_cache_*} API is high level, and uses a user level cache:\if{html}{\out{<div class="sourceCode r">}}\preformatted{pkg_cache_summary()
#> $cachepath
#> [1] "/Users/gaborcsardi/Library/Caches/org.R-project.R/R/pkgcache/pkg"
#> 
#> $files
#> [1] 696
#> 
#> $size
#> [1] 890853353
}\if{html}{\out{</div>}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{pkg_cache_list()
#> # A tibble: 696 x 11
#>    fullpath  path   package  url   etag  sha256 version platform built vignettes
#>    <chr>     <chr>  <chr>    <chr> <chr> <chr>  <chr>   <chr>    <chr> <chr>    
#>  1 /Users/g~ bin/m~ mockr    http~ "\"5~ e992e~ 0.1     macos    <NA>  <NA>     
#>  2 /Users/g~ bin/m~ dplyr    http~ "\"1~ 94269~ 1.0.6   macos    <NA>  <NA>     
#>  3 /Users/g~ bin/m~ waldo    http~ "\"1~ b7d0c~ 0.2.5   macos    <NA>  <NA>     
#>  4 /Users/g~ src/c~ downlit  <NA>   <NA> 0778f~ <NA>    <NA>     FALSE <NA>     
#>  5 /Users/g~ src/c~ cli      <NA>   <NA> a4577~ <NA>    <NA>     FALSE <NA>     
#>  6 /Users/g~ bin/m~ openssl  http~ "\"2~ 85b8b~ 1.4.4   macos    <NA>  <NA>     
#>  7 /Users/g~ bin/m~ tidysel~ http~ "\"3~ 4f76e~ 1.1.1   macos    <NA>  <NA>     
#>  8 /Users/g~ bin/m~ xfun     http~ "\"5~ 2f8bf~ 0.23    macos    <NA>  <NA>     
#>  9 /Users/g~ bin/m~ pillar   http~ "\"f~ 30aaa~ 1.6.1   macos    <NA>  <NA>     
#> 10 /Users/g~ bin/m~ tibble   http~ "\"c~ 10207~ 3.1.2   macos    <NA>  <NA>     
#> # ... with 686 more rows, and 1 more variable: rversion <chr>
}\if{html}{\out{</div>}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{pkg_cache_find(package = "dplyr")
#> # A tibble: 5 x 11
#>   fullpath   path   package url    etag  sha256 version platform built vignettes
#>   <chr>      <chr>  <chr>   <chr>  <chr> <chr>  <chr>   <chr>    <chr> <chr>    
#> 1 /Users/ga~ bin/m~ dplyr   https~ "\"1~ 94269~ 1.0.6   macos    <NA>  <NA>     
#> 2 /Users/ga~ bin/m~ dplyr   https~ "\"6~ c75d2~ 1.0.7   macos    <NA>  <NA>     
#> 3 /Users/ga~ bin/m~ dplyr   https~ "\"6~ c7fac~ 1.0.7   macos    <NA>  <NA>     
#> 4 /Users/ga~ bin/m~ dplyr   https~ "\"6~ c7fac~ 1.0.7   macos    <NA>  <NA>     
#> 5 /Users/ga~ bin/m~ dplyr   https~ "\"6~ c75d2~ 1.0.7   macos    <NA>  <NA>     
#> # ... with 1 more variable: rversion <chr>
}\if{html}{\out{</div>}}

\code{pkg_cache_add_file()} can be used to add a file,
\code{pkg_cache_delete_files()} to remove files, \code{pkg_cache_get_files()} to
copy files out of the cache.

The \code{package_cache} class provides a finer API.
}

\subsection{Bioconductor support}{

Both the metadata cache and the package cache support Bioconductor by
default, automatically. See the \code{BioC_mirror} option and the
\code{R_BIOC_MIRROR} and \code{R_BIOC_VERSION} environment variables below to
configure Bioconductor support.
}

\subsection{Package Options}{
\itemize{
\item The \code{BioC_mirror} option can be used to select a Bioconductor
mirror. This takes priority over the \code{R_BIOC_MIRROR} environment
variable.
\item \code{pkgcache_timeout} is the HTTP timeout for all downloads. It is in
seconds, and the limit for downloading the whole file. Defaults to
3600, one hour. It corresponds to the \href{https://curl.se/libcurl/c/CURLOPT_TIMEOUT.html}{\code{TIMEOUT} libcurl option}.
\item \code{pkgcache_connecttimeout} is the HTTP timeout for the connection
phase. It is in seconds and defaults to 30 seconds. It corresponds
to the \href{https://curl.se/libcurl/c/CURLOPT_CONNECTTIMEOUT.html}{\code{CONNECTTIMEOUT} libcurl option}.
\item \code{pkgcache_low_speed_limit} and \code{pkgcache_low_speed_time} are used
for a more sensible HTTP timeout. If the download speed is less than
\code{pkgcache_low_speed_limit} bytes per second for at least
\code{pkgcache_low_speed_time} seconds, the download errors. They
correspond to the
\href{https://curl.se/libcurl/c/CURLOPT_LOW_SPEED_LIMIT.html}{\code{LOW_SPEED_LIMIT}}
and
\href{https://curl.se/libcurl/c/CURLOPT_LOW_SPEED_TIME.html}{\code{LOW_SPEED_TIME}}
curl options.
}
}

\subsection{Package environment variables}{
\itemize{
\item The \code{R_BIOC_VERSION} environment variable can be used to override
the default Bioconductor version detection and force a given
version. E.g. this can be used to force the development version of
Bioconductor.
\item The \code{R_BIOC_MIRROR} environment variable can be used to select a
Bioconductor mirror. The \code{BioC_mirror} option takes priority over
this, if set.
\item \code{PKGCACHE_TIMEOUT} is the HTTP timeout for all downloads. It is in
seconds, and the limit for downloading the whole file. Defaults to
3600, one hour. It corresponds to the \href{https://curl.se/libcurl/c/CURLOPT_TIMEOUT.html}{\code{TIMEOUT} libcurl option}. The
\code{pkgcache_timeout} option has priority over this, if set.
\item \code{PKGCACHE_CONNECTTIMEOUT} is the HTTP timeout for the connection
phase. It is in seconds and defaults to 30 seconds. It corresponds
to the \href{https://curl.se/libcurl/c/CURLOPT_CONNECTTIMEOUT.html}{\code{CONNECTTIMEOUT} libcurl option}. The
\code{pkgcache_connecttimeout} option takes precedence over this, if set.
\item \code{PKGCACHE_LOW_SPEED_LIMIT} and \code{PKGCACHE_LOW_SPEED_TIME} are used
for a more sensible HTTP timeout. If the download speed is less than
\code{PKGCACHE_LOW_SPEED_LIMIT} bytes per second for at least
\code{PKGCACHE_LOW_SPEED_TIME} seconds, the download errors. They
correspond to the
\href{https://curl.se/libcurl/c/CURLOPT_LOW_SPEED_LIMIT.html}{\code{LOW_SPEED_LIMIT}}
and
\href{https://curl.se/libcurl/c/CURLOPT_LOW_SPEED_TIME.html}{\code{LOW_SPEED_TIME}}
curl options. The \code{pkgcache_low_speed_time} and
\code{pkgcache_low_speed_limit} options have priority over these
environment variables, if they are set.
\item \code{R_PKG_CACHE_DIR} is used for the cache directory, if set.
(Otherwise \code{rappdirs::user_cache_dir()} is used, see also
\code{meta_cache_summary()} and \code{pkg_cache_summary()}).
}
}

\subsection{Using pkgcache in CRAN packages}{

If you use pkgcache in your CRAN package, please make sure that
\itemize{
\item you don’t use pkgcache in your examples, and
\item you set the \code{R_USER_CACHE_DIR} environment variable to a temporary
directory (e.g. via \code{tempfile()}) during test cases. See the
\code{tests/testthat/setup.R} file in pkgcache for an example.
}

This is to make sure that pkgcache does not modify the user’s files
while running \verb{R CMD check}.
}

\subsection{Code of Conduct}{

Please note that the pkgcache project is released with a \href{https://github.com/r-lib/pkgcache/blob/master/.github/CODE_OF_CONDUCT.md}{Contributor Code of Conduct}.
By contributing to this project, you agree to abide by its terms.
}

\subsection{License}{

MIT (c) \href{https://www.rstudio.com/}{RStudio Inc}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/pkgcache#readme}
  \item Report bugs at \url{https://github.com/r-lib/pkgcache/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

}
